feat: Add intelligent test name preprocessing for AI referral matching
Implement Phase 2 improvements for AI test matching with preprocessing:

**New Features:**
1. Compound test splitting:
   - "Vit B12/Folate" → splits into B12 + FOL
   - "FBC+UEC+LFT" → splits into individual tests
   - Supports separators: /, &, and, +, comma

2. Abbreviation expansion:
   - "Vit" → "Vitamin"
   - "FBE" → "FBC"
   - "U&E" → "UEC"
   - Common Australian pathology abbreviations

3. Panel recognition:
   - "EIFT" → expands to UEC + IRON + FERR + TFT
   - "Cardiac Panel" → TROP + BNP + CK + CKMB
   - "Anemia Panel" → FBC + IRON + FERR + B12 + FOL

**Bug Fixes:**
- Fix fallback behavior: Return empty string for test_id when no match found
  (previously returned test name, causing frontend validation errors)
- Fix config validation: Change max_image_size_mb from int to float
  (allows 3.5MB limit for Claude Vision API)

**New Files:**
- src/app/services/test_preprocessor.py: Test name preprocessing logic

**Modified Files:**
- src/app/services/test_matcher.py: Integrate preprocessing, fix fallback
- src/app/config.py: Fix max_image_size_mb type validation

**Testing:**
Referral with "FBE, EIFT, TSH, Iron Studies, Vit D, Vit B12/Folate" now:
- Correctly expands to 9 tests (UEC, IRON, FERR, TFT from EIFT)
- Correctly splits B12/Folate into separate tests
- All aliases match correctly (FBE→FBC, Vit D→VITD)

+++ src/app/config.py
@@ -39,7 +39,7 @@ class Settings(BaseSettings):
     # Anthropic
     anthropic_api_key: str = ""
     anthropic_model: str = "claude-sonnet-4-5-20250929"
-    max_image_size_mb: int = 10
+    max_image_size_mb: float = 10.0  # Allow decimal precision for size limits
     scan_timeout_seconds: int = 120
 
     # External Services
